% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphas.R
\name{alphas}
\alias{alphas}
\title{Function to organize user-entered alpha levels.}
\usage{
alphas(a0, a1, a2)
}
\arguments{
\item{a0}{Up to 3 alpha levels.}

\item{a1}{Up to 3 alpha levels.}

\item{a2}{Up to 3 alpha levels.}
}
\value{
A numeric vector of confidence levels of length 3, ordered in increasing order with repetitions of the smallest value.
}
\description{
Duplicate values if needed to make 3. An error message is provided if there is not at least one valid level.
}
