% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawLegend.R
\name{drawlegend}
\alias{drawlegend}
\title{FUNCTION to draw legend and set up colors for regions for chart.}
\usage{
drawlegend(confLevel, MMEM, chartBW, xmax, ymax)
}
\arguments{
\item{confLevel}{Numeric 3-component vector of confidence levels with possible repeat values.}

\item{MMEM}{Minimum meaningful effect magnitude (after Fisher transformation in case of correlation).}

\item{chartBW}{TRUE if b&w chart is to be drawn.}

\item{xmax}{Extent of x-axis is -xmax to xmax. Used to position legend.}

\item{ymax}{Extent of y-axis is 0 to ymax, in units of SE. Used to position legend.}
}
\value{
Vector of colors (gray levels for b&w version). These cannot currently be changed by the user.
}
\description{
Start by setting up labels describing test levels.
Then set up labels to describe the tests; these differ if there is a minimum meaningful effect (MMEM>0).
Set colours and angles for regions according to whether black and white option selected and how many alpha levels.
If meaningful effects >0 there will be equivalence regions in the legend.
With minimum meaningful effects, there are non-superior and non-inferior regions as well as an inconclusive region.
}
