% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSSize.R
\name{genSSize}
\alias{genSSize}
\title{Launches Shiny app to calculate sample sizes with minimum meaningful effects and multiple tests and alpha levels.}
\usage{
genSSize(
  FunnelPlot = F,
  level = ".05, .005",
  power = 0.8,
  meaningfulEffect = 0,
  dataFile = "Cowlishaw.txt"
)
}
\arguments{
\item{FunnelPlot}{TRUE for a generalized funnel plot chart in meta-analysis rather than sample size calculator. Default FALSE.}

\item{level}{Comma-separated text containing between 1 and 3 test alpha levels. Default "0.05,0.005".}

\item{power}{Power desired in sample size calculation; ignored in funnel plot. Default 0.80.}

\item{meaningfulEffect}{Minimum meaningful effect magnitude. Default 0.}

\item{dataFile}{Text file name, used for meta-analyses only. File must be formatted as CSV with 2 rows, first row contains study means, second study variances.}
}
\description{
Tests for superiority, inferiority, non-superiority, non-inferiority and equivalence (with respect to a range of effect sizes equivalent to zero) are considered at up to 3 user-nominated alpha levels.
Currently, through the Shiny UI, users can select either a single group or a two group (equi-variance) design.
A chart displays effect size against total sample size. Colors relate regions in the chart to the strongest test that has the nominated power for an estimated effect size and sample size.
Users must nominate a power level and pooled variance as well as the alpha levels. Vertical lines at user-nominated effect show how changes in sample size affect plausible tests.
There is an option to produce black and white charts and users can change the scale of either axis.
\emph{Another option converts the chart to a generalized contour funnel plot which displays study and meta-analysis findings against rejection regions for tests with respect to minimum meaningful effects.}
}
\details{
INTERFACE:
Shiny side panel explains the sample size calculator and asks for information about the design (one or two group) and an approximate DOF for t tests.
The main panel starts by explaining the chart, which is displayed below. Beneath the chart are widgets to take user input.
The explanation and input options as well as the chart change if the option for a generalized funnel plot is selected.
Then the estimated effect size and variance are requested (or study means and vaiances if drawing a generalized funnel plot).
The desired power in the sample size computations is requested.
After that, display parameters are obtained. Currently these are the smallest sample size and the largest effect magnitude, plus whether the chart should be in black and white.
Finally users are advised of where to find details of the calculations, and offered the choice of showing a generalized funnel plot.
If this option is checked, the side panel disappears, and different input options are presented.
Either a fixed-effect or ML random-effects model can be selected, the calculations of which will use the CRAN metafor package.

DRAWING THE CHART:
To draw the chart, first setup parameters and draw an empty plot (function setup).
Next, draw the legend and work out the colors to use in drawing regions (these depend on the type of tests and number of levels) (function drawlegend).
Then draw the rejection regions for the various tests, which for the sample size calculations must also have sufficient power (function drawregions).
Next, display lines at the expected effect size(s) (or points representing study findings if funnel plot option selected) (function datadisplay).
Allow users to get co-ordinates of a point on click (function clickPoints).
Draw NHST boundaries if requested.
}
