% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPoints.R
\name{plotpoints}
\alias{plotpoints}
\title{FUNCTION to ensure corner points used to draw polygons (regions) in function drawregions aren't beyond chart boundaries.}
\usage{
plotpoints(xm, ym, MMEM, conf, power, nn)
}
\arguments{
\item{xm}{The x co-ordinate of the corner point of the polygon.}

\item{ym}{The y co-ordinate of the corner point of the polygon.}

\item{MMEM}{Minimum meaningful effect magnitude.}

\item{conf}{The test confidence level of this polygon.}

\item{power}{Desired power in sample size calculation.}

\item{nn}{DOF of t-test.}
}
\value{
Co-ordinates which don't extend beyond chart.
}
\description{
FUNCTION to ensure corner points used to draw polygons (regions) in function drawregions aren't beyond chart boundaries.
}
