% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUp.R
\name{setup}
\alias{setup}
\title{FUNCTION to set up parameters used in function genSSsize, and to draw empty plot.}
\usage{
setup(var, nn, MMEM, power, xmax, yhigh, ylow, FunnelPlot, Study)
}
\arguments{
\item{var}{Anticipated variance in sample size calculation. Used with Study parameter to convert sample sizes to SE.}

\item{nn}{DOF of t-test. If invalid will be set to default 38. Is re-set to 200 for funnel plot.}

\item{MMEM}{Minimum meaningful effect magnitude.}

\item{power}{Desired power in sample size calculation. Is re-set to 0 for funnel plot.}

\item{xmax}{Largest effect size to be displayed.}

\item{yhigh}{Largest SE to be displayed on funnel plot (smallest is 0).}

\item{ylow}{Smallest sample size to be displayed, if calculating sample size.}

\item{FunnelPlot}{TRUE if a generalized funnel plot is requested.}

\item{Study}{Numeric. One group (value 0 or 1) or t?setUp
wo group with value = relative size.}
}
\value{
Adjusted var, nn, MMEM, power, xmax, ymax (SE) or nmin (sample size), qtb (t-quantile),ytick (tick mark positions on y-axis).
}
\description{
First, set up technical parameters, giving feedback if input is invalid.
Use large DOF to simulate normal distributions for meta-analysis display.
Then set up visual display parameters.
y-axis will be in units of SE or sample size, from a nominated value to 0 or infinity respectively.
Labels for y axis will be total sample size unless generalized funnel plot requested.
Compute largest SE from smallest sample size if doing sample size calculation.
Use this to work out y-axis tick marks and labels in terms of sample size.
Draw empty plot with title that gives type of chart (sample size calculator or funnel plot).
Put tick marks on top of chart at smallest effect magnitudes.
}
